fix(sim): show fractional years/sec in window title

+++ opengl/sim/main.odin
@@ -104,7 +104,7 @@ update_window_title :: proc (window: glfw.WindowHandle, bodies: []Body) {
 		tracked_body_name := bodies[camera.tracked_body].name
 		title = fmt.ctprintf("%s - %s - %d days/sec - %d years/sec", "Sol_Sim", tracked_body_name, sim_speed / 86400, sim_speed / 3.156e7)
 	} else {
-		title = fmt.ctprintf("%s - %d days/sec - %d years/sec", "Sol_Sim", sim_speed / 86400, sim_speed / 3.156e7)
+		title = fmt.ctprintf("%s - %d days/sec - %f years/sec", "Sol_Sim", sim_speed / 86400, f64(sim_speed) / 3.156e7)
 	}
 
 	glfw.SetWindowTitle(window, title)
